Add -Wall/-Wextra/-Wpedantic via ray-tracer-flags target; fix all warnings
Introduce an INTERFACE target 'ray-tracer-flags' carrying the project warning
flags, linked only into ray-tracer-lib/ray-tracer/tests/editor (never Conan
deps). Add RAYTRACER_WERROR option. Build is warning-clean (144 -> 0) and
-Werror builds green.

Real fixes:
- Tree::getPivot primary stub returned a dangling reference to a stack temporary
  (-Wreturn-stack-address, UB if called) -> return function-local static.
- TypedObject: add virtual ~dtor; Light/Volume subtypes are created/destroyed
  through base handles via make_shared (-Wdelete-non-abstract-non-virtual-dtor, UB).
- PngWriter::setTitle: string-literal-to-png_charp conversion + reserve()/strcpy
  into zero-size vector -> sized zero-init vector + const_cast key.
- Bounds/Limits: default copy ctor (deprecated-copy); WorkQueue ctor init-order;
  Worker sign-compare loop; remove dead 'prev' vars; Buffer m_height maybe_unused.
- Triangle operator+ was 'static' in a header (46 per-TU unused-function) -> extern.
- Silence -Wpedantic anonymous-struct in Vector.h locally via pragma; suppress
  warnings on third-party imgui backend TUs.

+++ src/Worker.cpp
@@ -617,7 +617,7 @@ bool Worker::processPhotons()
             size_t minIndex = 0;
             bool validHit = false;
 
-            for (int i = 0; i < m_volumeHitBuffer.size(); ++i)
+            for (size_t i = 0; i < m_volumeHitBuffer.size(); ++i)
             {
                 if (m_volumeHitBuffer[i].hit.distance < minDistance && m_volumeHitBuffer[i].hit.distance > selfHitThreshold)
                 {
